# proxy/docs/swagger.yaml
basePath: /api
definitions:
  internal_controller_http_v1.RequestAddressGeocode:
    properties:
      lat:
        example: "55.753214"
        type: string
      lng:
        example: "37.642589"
        type: string
    required:
    - lat
    - lng
    type: object
  internal_controller_http_v1.RequestAddressSearch:
    properties:
      query:
        example: Подкопаевский переулок
        type: string
    required:
    - query
    type: object
  rr.JSONResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Geoservice API
  title: Microservice Geoservice
  version: 2.0.0
paths:
  /address/geocode:
    post:
      consumes:
      - application/json
      parameters:
      - description: coordinates
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/internal_controller_http_v1.RequestAddressGeocode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rr.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rr.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Returns a list of addresses provided geo coordinates
      tags:
      - address
  /address/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: street name
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/internal_controller_http_v1.RequestAddressSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rr.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rr.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Returns a list of addresses provided street name
      tags:
      - address
  /auth/login:
    post:
      parameters:
      - description: Email for login (john.doe@gmail.com)
        in: formData
        name: email
        required: true
        type: string
      - description: Password for login (password)
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rr.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rr.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rr.JSONResponse'
      summary: Logs user into the system
      tags:
      - auth
  /auth/logout:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rr.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rr.JSONResponse'
      summary: Logs out current user
      tags:
      - auth
  /auth/register:
    post:
      parameters:
      - description: New user email
        in: formData
        name: email
        required: true
        type: string
      - description: New user password
        in: formData
        name: password
        required: true
        type: string
      - description: User first name
        in: formData
        name: firstName
        type: string
      - description: User last name
        in: formData
        name: lastName
        type: string
      - description: User age in years
        in: formData
        name: age
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rr.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rr.JSONResponse'
      summary: Creates new user
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"